feat: User creation on custom Adapter
The adapter of user creation, is important to see that for some reason
NextAuth is not typing on UserAdapter type the expires_at as Datetime,
is using number. Other important thing is that Google dont provide the
avatar_url as avatar_url and with "picture".
After the adapter was made some refactoring in the home page for better
behaviour of the communication between home and register page.