feat: all 4 steps to register is done
The last step to update the bio and  the other steps connection to make
the behaviour of next step is completed.

Important note: If you have already a user that is authenticated with
Google account, when try to make a new user with the same account the
auth will think that the changes is realated to the first user created,
because the id of the google account is linked to the first created
user.