feat: blocked times
Blocked times based on the already scheduled times, and blocked times
because is later than the possible time.
Experience with times and querys, good thing the remember is that
javascript start with january being month 0 and by side that on query is
needed the 0number like 05 an the js was passing 5.